Fix clang-tidy bugprone-implicit-widening-of-multiplicati… (#2318)
*  Fix clang-tidy bugprone-implicit-widening-of-multiplication-result check

* Update pytorch_helpers/habana_helpers/logging.cpp


---------

+++ habana_eager/strided_dynamic_ops.cpp
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2024 Intel Corporation
+ * Copyright (c) 2021-2025 Intel Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -449,7 +449,7 @@ bool ConstantPad2dOperatorDS::ReplaceWithDynamicHPUOp(
   std::vector<std::string> pad_ht_vec_expr(MAX_DIMENSIONS_NUM * 2, "0");
   // assuming that "pad" has a pair of pad values corresponding to each
   // dim that needs to be padded.
-  for (unsigned int i = 0; i < values.size() / 2; i++) {
+  for (size_t i = 0; i < values.size() / 2; i++) {
     // Host tensor layout 1D - 10 elements:
     // pad_before[0]...pad_before[4], pad_after[0] ... pad_after[4] (for
     // dimensionality IFM less then 5 some elements not in use)
