+++ habana_kernels/embedding_kernels.cpp
@@ -61,7 +61,7 @@ std::vector<int64_t> PadOperator::compute_output_shape(
 
   auto shape = self.sizes().vec();
 
-  for (unsigned int i = 0; i < lpad; i++) {
+  for (size_t i = 0; i < lpad; i++) {
     auto pad_start = pad[2 * i];
     auto pad_end = pad[2 * i + 1];
     shape[ndim - i - 1] += (pad_start + pad_end);
@@ -143,7 +143,7 @@ void PadOperator::AllocateAndAddSynapseNode(
     param.value.f = inputs[2].toScalar().to<float>();
   }
   memset(param.pads, 0, sizeof(param.pads));
-  for (unsigned int i = 0; i < lpad; i++) {
+  for (size_t i = 0; i < lpad; i++) {
     param.pads[i] = pad[2 * i];
     param.pads[i + ndim] = pad[2 * i + 1];
   }
